pyproject.toml
[project]
name = "entropy-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.56.0",
    "groq>=1.0.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.30",
    "uvicorn[standard]>=0.40.0",
]
